requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1
python-dotenv==1.0.0
websockets==11.0.3
requests==2.31.0
pydantic==2.4.2
jinja2==3.1.2
aiofiles==23.2.1